+++ workspaces/frontend/src/app/pages/Workspaces/Form/properties/volumes/VolumesAttachModal.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react';
+import React, { useCallback, useEffect, useMemo, useState } from 'react';
 import { Button } from '@patternfly/react-core/dist/esm/components/Button';
 import {
   Modal,
@@ -10,56 +10,23 @@ import {
 import { Alert, AlertVariant } from '@patternfly/react-core/dist/esm/components/Alert';
 import { Form, FormGroup } from '@patternfly/react-core/dist/esm/components/Form';
 import { Switch } from '@patternfly/react-core/dist/esm/components/Switch';
-import { Label, LabelGroup } from '@patternfly/react-core/dist/esm/components/Label';
-import { Flex, FlexItem } from '@patternfly/react-core/dist/esm/layouts/Flex';
 import { Stack, StackItem } from '@patternfly/react-core/dist/esm/layouts/Stack';
-import { Tooltip } from '@patternfly/react-core/dist/esm/components/Tooltip';
-import { InfoCircleIcon } from '@patternfly/react-icons/dist/esm/icons/info-circle-icon';
-import { CubeIcon } from '@patternfly/react-icons/dist/esm/icons/cube-icon';
-import { TimesIcon } from '@patternfly/react-icons/dist/esm/icons/times-icon';
-import { Divider } from '@patternfly/react-core/dist/esm/components/Divider';
-import {
-  Select,
-  SelectGroup,
-  SelectList,
-  SelectOption,
-} from '@patternfly/react-core/dist/esm/components/Select';
-import { MenuToggle } from '@patternfly/react-core/dist/esm/components/MenuToggle';
-import {
-  TextInputGroup,
-  TextInputGroupMain,
-  TextInputGroupUtilities,
-} from '@patternfly/react-core/dist/esm/components/TextInputGroup';
+import { TypeaheadSelect } from '@patternfly/react-templates';
 import {
   PvcsPVCListItem,
   StorageclassesStorageClassListItem,
   V1PersistentVolumeAccessMode,
 } from '~/generated/data-contracts';
 import ThemeAwareFormGroupWrapper from '~/shared/components/ThemeAwareFormGroupWrapper';
-import { LabelGroupWithTooltip } from '~/app/components/LabelGroupWithTooltip';
 import {
   normalizeMountPath,
   validateMountPath,
   getMountPathUniquenessError,
+  buildPVCSelectOptions,
 } from '~/app/pages/Workspaces/Form/helpers';
 import { MountPathField } from '~/app/pages/Workspaces/Form/MountPathField';
 
-interface PVCOptionData {
-  /** Display text and filter target */
-  content: string;
-  value: string;
-  isDisabled: boolean;
-  description: React.ReactNode;
-  /** Explains why the PVC is unmountable, shown as a tooltip on hover */
-  tooltip: string | null;
-}
-
-interface PVCGroupData {
-  label: string;
-  displayName: string;
-  description: string;
-  options: PVCOptionData[];
-}
+const PVC_SELECT_EMPTY_KEY = 'pvc-select-empty';
 
 export interface VolumesAttachModalProps {
   isOpen: boolean;
@@ -85,20 +52,6 @@ const isRWO = (pvc: PvcsPVCListItem): boolean =>
 
 const isInUse = (pvc: PvcsPVCListItem): boolean => pvc.pods.length > 0 || pvc.workspaces.length > 0;
 
-const getUnmountableTooltip = (pvc: PvcsPVCListItem): string | null => {
-  if (pvc.canMount) {
-    return null;
-  }
-  const modes = pvc.pvcSpec.accessModes;
-  if (modes.includes(V1PersistentVolumeAccessMode.ReadWriteOncePod) && pvc.pods.length > 0) {
-    return 'This volume uses ReadWriteOncePod access and is already mounted by a pod.';
-  }
-  if (modes.includes(V1PersistentVolumeAccessMode.ReadWriteOnce) && pvc.workspaces.length > 0) {
-    return 'This volume uses ReadWriteOnce access and is already mounted by a workspace.';
-  }
-  return null;
-};
-
 export const VolumesAttachModal: React.FC<VolumesAttachModalProps> = ({
   isOpen,
   setIsOpen,
@@ -109,7 +62,8 @@ export const VolumesAttachModal: React.FC<VolumesAttachModalProps> = ({
   excludedPvcNames,
   storageClasses,
 }) => {
-  // Form state
+  // ── Form state ───────────────────────────────────────────────────────────
+
   const [selectedPvcName, setSelectedPvcName] = useState('');
   const [mountPath, setMountPath] = useState(fixedMountPath ?? '/data/');
   const [isMountPathEditing, setIsMountPathEditing] = useState(false);
@@ -122,13 +76,7 @@ export const VolumesAttachModal: React.FC<VolumesAttachModalProps> = ({
     : null;
   const mountPathError = mountPathFormatError ?? mountPathUniquenessError;
 
-  // Typeahead select state
-  const [isSelectOpen, setIsSelectOpen] = useState(false);
-  const [inputValue, setInputValue] = useState('');
-  const [filterValue, setFilterValue] = useState('');
-  const [focusedItemIndex, setFocusedItemIndex] = useState<number | null>(null);
-  const [activeItemId, setActiveItemId] = useState<string | null>(null);
-  const textInputRef = useRef<HTMLInputElement>(undefined);
+  // ── Reset on open ────────────────────────────────────────────────────────
 
   useEffect(() => {
     if (isOpen) {
@@ -137,13 +85,10 @@ export const VolumesAttachModal: React.FC<VolumesAttachModalProps> = ({
       setIsMountPathEditing(false);
       setReadOnly(false);
       setFormError('');
-      setIsSelectOpen(false);
-      setInputValue('');
-      setFilterValue('');
-      setFocusedItemIndex(null);
-      setActiveItemId(null);
     }
   }, [isOpen, fixedMountPath]);
+
+  // ── Mount path handlers ──────────────────────────────────────────────────
 
   const handleStartMountPathEdit = useCallback(() => {
     setIsMountPathEditing(true);
@@ -167,266 +112,6 @@ export const VolumesAttachModal: React.FC<VolumesAttachModalProps> = ({
     }
     setIsMountPathEditing(false);
   }, [selectedPvcName, fixedMountPath]);
-
-  // ── PVC option building ──────────────────────────────────────────────────
-
-  const buildDescription = useCallback(
-    (pvc: PvcsPVCListItem): React.ReactNode => {
-      const isExcluded = excludedPvcNames?.has(pvc.name) ?? false;
-      return (
-        <Flex justifyContent={{ default: 'justifyContentSpaceBetween' }}>
-          <FlexItem>
-            <Stack style={{ width: '100%' }}>
-              <StackItem>
-                <LabelGroup numLabels={5}>
-                  <Label
-                    isCompact
-                    className={!pvc.canMount || isExcluded ? 'pf-m-disabled' : undefined}
-                  >
-                    {pvc.pvcSpec.requests.storage}
-                  </Label>
-                  {pvc.pvcSpec.accessModes.map((mode) => (
-                    <Label
-                      key={mode}
-                      isCompact
-                      className={!pvc.canMount || isExcluded ? 'pf-m-disabled' : undefined}
-                      color="blue"
-                    >
-                      {mode}
-                    </Label>
-                  ))}
-                  {isExcluded && (
-                    <Label isCompact className="pf-m-disabled">
-                      Already mounted
-                    </Label>
-                  )}
-                  {!pvc.canMount && (
-                    <Label isCompact className="pf-m-disabled">
-                      Unmountable
-                    </Label>
-                  )}
-                </LabelGroup>
-              </StackItem>
-              {pvc.workspaces.length > 0 && (
-                <StackItem className="pf-v6-u-ml-sm pf-v6-u-mt-xs">
-                  <Flex gap={{ default: 'gapXs' }}>
-                    <FlexItem>Connected Workspaces:</FlexItem>
-                    <FlexItem>
-                      <LabelGroupWithTooltip
-                        labels={pvc.workspaces.map((w) => w.name)}
-                        limit={5}
-                        variant="outline"
-                        icon={<CubeIcon color="teal" />}
-                        isCompact
-                        color="teal"
-                        className={!pvc.canMount ? 'pf-m-disabled' : undefined}
-                      />
-                    </FlexItem>
-                  </Flex>
-                </StackItem>
-              )}
-            </Stack>
-          </FlexItem>
-          <FlexItem>
-            <Tooltip
-              aria="none"
-              aria-live="polite"
-              content={
-                <Stack>
-                  <StackItem>{`Created at: ${new Date(pvc.audit.createdAt).toLocaleString()} by ${pvc.audit.createdBy}`}</StackItem>
-                  <StackItem>{`Updated at: ${new Date(pvc.audit.updatedAt).toLocaleString()} by ${pvc.audit.updatedBy}`}</StackItem>
-                </Stack>
-              }
-            >
-              <span style={{ cursor: 'default' }}>
-                <InfoCircleIcon />
-              </span>
-            </Tooltip>
-          </FlexItem>
-        </Flex>
-      );
-    },
-    [excludedPvcNames],
-  );
-
-  // ── Grouped / filtered option data ──────────────────────────────────────
-
-  const pvcGroups = useMemo((): PVCGroupData[] => {
-    const scMap = new Map(storageClasses.map((s) => [s.name, s]));
-    const grouped = new Map<
-      string,
-      { description: string; displayName: string; options: PVCOptionData[] }
-    >();
-    for (const pvc of availablePVCs) {
-      const sc = pvc.pvcSpec.storageClassName || 'default';
-      if (!grouped.has(sc)) {
-        grouped.set(sc, {
-          description: scMap.get(sc)?.description ?? '',
-          displayName: scMap.get(sc)?.displayName ?? sc,
-          options: [],
-        });
-      }
-      grouped.get(sc)!.options.push({
-        content: pvc.name,
-        value: pvc.name,
-        isDisabled: !pvc.canMount || (excludedPvcNames?.has(pvc.name) ?? false),
-        description: buildDescription(pvc),
-        tooltip: getUnmountableTooltip(pvc),
-      });
-    }
-    return Array.from(grouped.entries()).map(([sc, { description, displayName, options }]) => ({
-      label: displayName || sc,
-      displayName,
-      description,
-      options: options.sort((a, b) => Number(a.isDisabled) - Number(b.isDisabled)),
-    }));
-  }, [availablePVCs, buildDescription, excludedPvcNames, storageClasses]);
-
-  /** Flat ordered list used for keyboard navigation and focus tracking. */
-  const filteredFlatOptions = useMemo((): PVCOptionData[] => {
-    const all = pvcGroups.flatMap((g) => g.options);
-    if (!filterValue) {
-      return all;
-    }
-    return all.filter((opt) => opt.content.toLowerCase().includes(filterValue.toLowerCase()));
-  }, [pvcGroups, filterValue]);
-
-  /** Groups with per-group filtering applied; empty groups are excluded. */
-  const filteredGroups = useMemo((): PVCGroupData[] => {
-    if (!filterValue) {
-      return pvcGroups;
-    }
-    return pvcGroups
-      .map((g) => ({
-        ...g,
-        options: g.options.filter((opt) =>
-          opt.content.toLowerCase().includes(filterValue.toLowerCase()),
-        ),
-      }))
-      .filter((g) => g.options.length > 0);
-  }, [pvcGroups, filterValue]);
-
-  // ── Typeahead select handlers ────────────────────────────────────────────
-
-  const openMenu = useCallback(() => {
-    setIsSelectOpen(true);
-  }, []);
-
-  const closeMenu = useCallback(() => {
-    setIsSelectOpen(false);
-    setFocusedItemIndex(null);
-    setActiveItemId(null);
-  }, []);
-
-  const handleSelectOption = useCallback(
-    (value: string) => {
-      const pvcName = value;
-      setSelectedPvcName(pvcName);
-      setMountPath(fixedMountPath ?? `/data/${pvcName}`);
-      setIsMountPathEditing(false);
-      setInputValue(pvcName);
-      setFilterValue('');
-      setFormError('');
-      closeMenu();
-    },
-    [closeMenu, fixedMountPath],
-  );
-
-  const handleInternalSelect = useCallback(
-    (_ev: React.MouseEvent | undefined, value?: string | number) => {
-      if (value === undefined) {
-        return;
-      }
-      const opt = filteredFlatOptions.find((o) => o.value === String(value));
-      if (opt?.isDisabled) {
-        return;
-      }
-      handleSelectOption(String(value));
-    },
-    [handleSelectOption, filteredFlatOptions],
-  );
-
-  const handleInputChange = useCallback(
-    (_ev: React.FormEvent<HTMLInputElement>, value: string) => {
-      setInputValue(value);
-      setFilterValue(value);
-      setFocusedItemIndex(null);
-      setActiveItemId(null);
-      if (!isSelectOpen) {
-        openMenu();
-      }
-    },
-    [isSelectOpen, openMenu],
-  );
-
-  const handleInputClick = useCallback(() => {
-    if (!isSelectOpen) {
-      openMenu();
-    } else if (!inputValue) {
-      closeMenu();
-    }
-  }, [isSelectOpen, inputValue, openMenu, closeMenu]);
-
-  const handleToggleClick = useCallback(() => {
-    setIsSelectOpen((prev) => !prev);
-    textInputRef.current?.focus();
-  }, []);
-
-  const handleClearButtonClick = useCallback(() => {
-    setSelectedPvcName('');
-    setInputValue('');
-    setFilterValue('');
-    setFocusedItemIndex(null);
-    setActiveItemId(null);
-    textInputRef.current?.focus();
-  }, []);
-
-  const handleKeyDown = useCallback(
-    (event: React.KeyboardEvent<HTMLInputElement>) => {
-      const total = filteredFlatOptions.length;
-      if (total === 0) {
-        return;
-      }
-
-      if (event.key === 'Enter') {
-        event.preventDefault();
-        if (isSelectOpen && focusedItemIndex !== null) {
-          const focused = filteredFlatOptions[focusedItemIndex];
-          if (!focused.isDisabled) {
-            handleSelectOption(focused.value);
-          }
-        } else {
-          openMenu();
-        }
-        return;
-      }
-
-      if (event.key !== 'ArrowUp' && event.key !== 'ArrowDown') {
-        return;
-      }
-      event.preventDefault();
-      openMenu();
-
-      let next = 0;
-      if (event.key === 'ArrowUp') {
-        next =
-          focusedItemIndex === null || focusedItemIndex === 0 ? total - 1 : focusedItemIndex - 1;
-        while (filteredFlatOptions[next].isDisabled) {
-          next = next === 0 ? total - 1 : next - 1;
-        }
-      } else {
-        next =
-          focusedItemIndex === null || focusedItemIndex === total - 1 ? 0 : focusedItemIndex + 1;
-        while (filteredFlatOptions[next].isDisabled) {
-          next = next === total - 1 ? 0 : next + 1;
-        }
-      }
-
-      setFocusedItemIndex(next);
-      setActiveItemId(filteredFlatOptions[next].value);
-    },
-    [filteredFlatOptions, focusedItemIndex, isSelectOpen, openMenu, handleSelectOption],
-  );
 
   // ── In-use warning alert ─────────────────────────────────────────────────
 
@@ -467,42 +152,14 @@ export const VolumesAttachModal: React.FC<VolumesAttachModalProps> = ({
     onAttach(selectedPvc, trimmedPath, readOnly);
   }, [selectedPvc, mountPath, readOnly, mountedPaths, onAttach]);
 
-  // ── Render ───────────────────────────────────────────────────────────────
+  // ── Options ──────────────────────────────────────────────────────────────
 
-  const toggle = (toggleRef: React.Ref<HTMLButtonElement>) => (
-    <MenuToggle
-      ref={toggleRef}
-      variant="typeahead"
-      aria-label="Volume typeahead menu toggle"
-      onClick={handleToggleClick}
-      isExpanded={isSelectOpen}
-      isFullWidth
-    >
-      <TextInputGroup isPlain>
-        <TextInputGroupMain
-          value={inputValue}
-          onClick={handleInputClick}
-          onChange={handleInputChange}
-          onKeyDown={handleKeyDown}
-          autoComplete="off"
-          innerRef={textInputRef}
-          placeholder="Select a volume"
-          role="combobox"
-          isExpanded={isSelectOpen}
-          aria-controls="pvc-select-listbox"
-          {...(activeItemId ? { 'aria-activedescendant': activeItemId } : {})}
-        />
-        <TextInputGroupUtilities {...(!inputValue ? { style: { display: 'none' } } : {})}>
-          <Button
-            variant="plain"
-            onClick={handleClearButtonClick}
-            aria-label="Clear volume selection"
-            icon={<TimesIcon />}
-          />
-        </TextInputGroupUtilities>
-      </TextInputGroup>
-    </MenuToggle>
+  const initialOptions = useMemo(
+    () => buildPVCSelectOptions(availablePVCs, storageClasses, excludedPvcNames, selectedPvcName),
+    [availablePVCs, storageClasses, excludedPvcNames, selectedPvcName],
   );
+
+  // ── Render ───────────────────────────────────────────────────────────────
 
   return (
     <Modal
@@ -533,64 +190,22 @@ export const VolumesAttachModal: React.FC<VolumesAttachModalProps> = ({
           <StackItem>
             <Form>
               <ThemeAwareFormGroupWrapper label="Volume" fieldId="pvc-select">
-                <Select
+                <TypeaheadSelect
+                  key={selectedPvcName || PVC_SELECT_EMPTY_KEY}
                   id="pvc-select"
-                  isOpen={isSelectOpen}
-                  selected={selectedPvcName}
-                  onSelect={handleInternalSelect}
-                  onOpenChange={(open) => {
-                    if (!open) {
-                      closeMenu();
-                    }
-                  }}
-                  toggle={toggle}
-                  variant="typeahead"
+                  initialOptions={initialOptions}
+                  placeholder="Select a volume"
                   isScrollable
                   maxMenuHeight="25rem"
-                >
-                  {filteredGroups.length > 0 ? (
-                    filteredGroups.map((group, index) => (
-                      <React.Fragment key={group.label}>
-                        {index > 0 && <Divider />}
-                        <SelectGroup
-                          label={`${group.displayName ? group.displayName : group.label} ${group.description ? `- ${group.description}` : ''}`}
-                        >
-                          <SelectList>
-                            {group.options.map((opt) => {
-                              const flatIndex = filteredFlatOptions.findIndex(
-                                (o) => o.value === opt.value,
-                              );
-                              return (
-                                <SelectOption
-                                  key={opt.value}
-                                  value={opt.value}
-                                  isDisabled={opt.isDisabled && !opt.tooltip}
-                                  isAriaDisabled={opt.isDisabled && !!opt.tooltip}
-                                  {...(opt.tooltip
-                                    ? { tooltipProps: { content: opt.tooltip } }
-                                    : {})}
-                                  isFocused={focusedItemIndex === flatIndex}
-                                  id={opt.value}
-                                  description={opt.description}
-                                >
-                                  {opt.content}
-                                </SelectOption>
-                              );
-                            })}
-                          </SelectList>
-                        </SelectGroup>
-                      </React.Fragment>
-                    ))
-                  ) : (
-                    <SelectList>
-                      <SelectOption isAriaDisabled value="no-results">
-                        {filterValue
-                          ? `No volume found for "${filterValue}"`
-                          : 'No volumes available'}
-                      </SelectOption>
-                    </SelectList>
-                  )}
-                </Select>
+                  noOptionsFoundMessage={(filter) => `No volume found for "${filter}"`}
+                  onSelect={(_ev, value) => {
+                    setSelectedPvcName(value as string);
+                    setMountPath(fixedMountPath ?? `/data/${value}`);
+                    setIsMountPathEditing(false);
+                    setFormError('');
+                  }}
+                  onClearSelection={() => setSelectedPvcName('')}
+                />
               </ThemeAwareFormGroupWrapper>
               <MountPathField
                 variant="input"
